Cleanup implementation of graphql queries/mutations
Improved naming and code organization of how queries and mutations were being done on the client side + moved Admin client stuff to api-lib
