+++ src/lib/gql/queries.ts
@@ -1,0 +1,21 @@
+import { client } from './client';
+
+export const getCurrentEpoch = async (
+  circle_id: number
+): Promise<typeof currentEpoch | undefined> => {
+  const {
+    epochs: [currentEpoch],
+  } = await client.query({
+    epochs: [
+      {
+        where: {
+          circle_id: { _eq: circle_id },
+          end_date: { _gt: 'now()' },
+          start_date: { _lt: 'now()' },
+        },
+      },
+      { id: true },
+    ],
+  });
+  return currentEpoch;
+};
