+++ src/pages/DistributePage/useGetAllocations.ts
@@ -7,7 +7,7 @@ export function useGetAllocations(epochId: number) {
   // from: epoch -> circle -> users -> gifts
   // to: epoch -> gifts -> users
   // that would remove the need to repeatedly pass epochId as an argument
-  return useTypedQuery(`circle-for-epoch-${epochId}`, {
+  return useTypedQuery(['circle-for-epoch', epochId], {
     epochs_by_pk: [
       { id: epochId },
       {
